+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added config seeds for avoiding manual process with seeds ([#6](https://github.com/8xFF/atm0s-small-p2p/pull/6))
 - switch to use peer_id for send and recv ([#5](https://github.com/8xFF/atm0s-small-p2p/pull/5))
 - visualization service ([#3](https://github.com/8xFF/atm0s-small-p2p/pull/3))
-- mininum working version with service pkt and stream
+- minimum working version with service pkt and stream
 
 ### Fixed